implement toggle style

+++ game.h
@@ -29,7 +29,7 @@ void game_print(Game *game, char *str);
 bool game_can_move(Game *game);
 void game_set_prob(Game *game, double prob);
 double game_get_prob(Game *game);
-double game_get_prob_next(Game *game);
+int game_get_prob_next(Game *game);
 int game_count(Game *game);
 
 #endif
